Update Admin.js
Just a few edits as seen. Overall this is very well written. Overall logic makes sense  and I am able to follow the code well. The only suggestion I have is placing braces on newlines to help with the cleanliness of the code. One other recommendation I have is to write more comments if possible. Even though the code is very well written it is good to document a lot. Also put error message in all caps just to make it more visible to user. Might want to consider making it larger as well.